Added checks to see if dependencies are installed before installing

+++ scripts/installdeps.sh
@@ -16,7 +16,21 @@
 
 # initOS discovers the operating system for this system.
 
+[[ $DEBUG ]] || DEBUG='false'
+
 MOVE2KUBE_DEP_INSTALL_PATH="$PWD/bin"
+
+HAS_DOCKER="$(type docker &>/dev/null && echo true || echo false)"
+HAS_PACK="$(type pack &>/dev/null && echo true || echo false)"
+HAS_KUBECTL="$(type kubectl &>/dev/null && echo true || echo false)"
+HAS_OPERATOR_SDK="$(type operator-sdk &>/dev/null && echo true || echo false)"
+if [ "$HAS_OPERATOR_SDK" == 'true' ]; then
+    if operator-sdk version | cut -d, -f1 | grep 'operator-sdk version: "v1' -q; then
+        OPERATOR_SDK_V1='true'
+    else
+        OPERATOR_SDK_V1='false'
+    fi
+fi
 
 initOS() {
     OS="$(uname | tr '[:upper:]' '[:lower:]')"
@@ -35,16 +49,44 @@ askBeforeProceeding() {
     read -r -p 'Proceed? [y/N]:' oktoproceed
 }
 
+# cleanup temporary files.
+cleanup() {
+    if [[ -d "${MOVE2KUBE_DEP_INSTALL_PATH:-}" ]]; then
+        rm -rf "$MOVE2KUBE_DEP_INSTALL_PATH"
+    fi
+}
+
+# fail_trap is executed if an error occurs.
+fail_trap() {
+    result=$?
+    if [ "$result" != "0" ]; then
+        echo "Failed to install the dependencies."
+        echo -e "\tFor support, go to https://github.com/konveyor/move2kube"
+    fi
+    cleanup
+    exit $result
+}
+
 # MAIN
+
+#Stop execution on any error
+trap "fail_trap" EXIT
+set -e
+set -u
+
+# Set debug if desired
+if [ "${DEBUG}" == "true" ]; then
+    set -x
+fi
 
 initOS
 
-if [ "$OS" == "darwin" ]; then
+if [ "$OS" == 'darwin' ]; then
     PACKPLATFORM='macos'
     KUBECTLPLATFORM='darwin'
     OPERATORSDKPLATFORM='apple-darwin'
     echo 'Once this installation finishes, please do install docker from https://docs.docker.com/docker-for-mac/install/'
-elif [ "$OS" == "linux" ]; then
+elif [ "$OS" == 'linux' ]; then
     PACKPLATFORM='linux'
     KUBECTLPLATFORM='linux'
     OPERATORSDKPLATFORM='linux-gnu'
@@ -60,18 +102,36 @@ if [ "$oktoproceed" != 'y' ]; then
     exit 1
 fi
 
-if [ "$OS" == "linux" ]; then
+mkdir -p "$MOVE2KUBE_DEP_INSTALL_PATH"
+
+if [ "${HAS_DOCKER}" != 'true' ] && [ "$OS" == 'linux' ]; then
     if ! grep -q container_t '/proc/1/attr/current'; then
+        echo 'Installing docker...'
         curl -fsSL 'https://get.docker.com' -o 'get-docker.sh' && sudo sh 'get-docker.sh' && rm 'get-docker.sh'
+        echo 'Done.'
     fi
 fi
 
-mkdir -p "$MOVE2KUBE_DEP_INSTALL_PATH"
-curl -o pack.tgz -LJO 'https://github.com/buildpacks/pack/releases/download/v0.12.0/pack-v0.12.0-'"$PACKPLATFORM"'.tgz' && tar -xzf pack.tgz && mv pack "$MOVE2KUBE_DEP_INSTALL_PATH" && rm pack.tgz
-curl -o kubectl -LJO 'https://storage.googleapis.com/kubernetes-release/release/'"$(curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt)"'/bin/'"$KUBECTLPLATFORM"'/amd64/kubectl' && mv kubectl "$MOVE2KUBE_DEP_INSTALL_PATH"
-chmod +x "$MOVE2KUBE_DEP_INSTALL_PATH"/kubectl
-curl -o operator-sdk -LJO 'https://github.com/operator-framework/operator-sdk/releases/download/v1.0.0/operator-sdk-v1.0.0-x86_64-'"$OPERATORSDKPLATFORM" && mv operator-sdk "$MOVE2KUBE_DEP_INSTALL_PATH"
-chmod +x "$MOVE2KUBE_DEP_INSTALL_PATH"/operator-sdk
+if [ "${HAS_PACK}" != 'true' ]; then
+    echo 'Installing pack...'
+    curl -o pack.tgz -LJO 'https://github.com/buildpacks/pack/releases/download/v0.12.0/pack-v0.12.0-'"$PACKPLATFORM"'.tgz' && tar -xzf pack.tgz && mv pack "$MOVE2KUBE_DEP_INSTALL_PATH" && rm pack.tgz
+    echo 'Done.'
+fi
+
+if [ "${HAS_KUBECTL}" != 'true' ]; then
+    echo 'Installing kubectl...'
+    curl -o kubectl -LJO 'https://storage.googleapis.com/kubernetes-release/release/'"$(curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt)"'/bin/'"$KUBECTLPLATFORM"'/amd64/kubectl' && mv kubectl "$MOVE2KUBE_DEP_INSTALL_PATH"
+    chmod +x "$MOVE2KUBE_DEP_INSTALL_PATH"/kubectl
+    echo 'Done.'
+fi
+
+if [ "${HAS_OPERATOR_SDK}" != 'true' ] || [ "$OPERATOR_SDK_V1" != 'true' ]; then
+    echo 'Installing operator-sdk...'
+    curl -o operator-sdk -LJO 'https://github.com/operator-framework/operator-sdk/releases/download/v1.0.0/operator-sdk-v1.0.0-x86_64-'"$OPERATORSDKPLATFORM" && mv operator-sdk "$MOVE2KUBE_DEP_INSTALL_PATH"
+    chmod +x "$MOVE2KUBE_DEP_INSTALL_PATH"/operator-sdk
+    echo 'Done.'
+fi
+
 echo 'PATH="$PATH:'"$MOVE2KUBE_DEP_INSTALL_PATH"'"' >>~/.bash_profile
 echo 'Installed the dependencies to '"$MOVE2KUBE_DEP_INSTALL_PATH"
 echo 'We have added a line to your '~/.bash_profile' to put '"$MOVE2KUBE_DEP_INSTALL_PATH"' on your $PATH. Either restart the shell or source ~/.bash_profile to see the changes.'
